settings/requirements.txt
Django==3.2.9
djangorestframework==3.13.1
drf_yasg==1.20.0
Faker==9.8.1
